Expose feature-block Arrow conversion as packaged s2and.arrow_service_io
Services that install s2and (without the scripts/ tree) need to convert an
in-memory ANDData into the Arrow IPC artifacts that predict_from_arrow_paths
consumes. Move write_feature_block_arrow_from_anddata + feature_block_from_anddata
into the packaged s2and.arrow_service_io; scripts/arrow_conversion_helpers.py
re-exports for backwards compatibility.
